moved string util functions to dedicated header
