+++ data/org.containers.hirte.Monitor.xml
@@ -1,7 +1,7 @@
 <!DOCTYPE node PUBLIC "-//freedesktop//DTD D-BUS Object Introspection 1.0//EN" "http://www.freedesktop.org/standards/dbus/1.0/introspect.dtd">
 <!-- SPDX-License-Identifier: GPL-2.0-or-later -->
 <node>
-  <interface name="org.freedesktop.systemd1.Manager">
+  <interface name="org.containers.hirte.Monitor">
     <method name="Close" />
     <method name="Subscribe">
       <arg name="node" type="s" direction="in" />
