+++ src/libhirte/common/common.h
@@ -10,20 +10,7 @@
 
 #include "list.h"
 #include "protocol.h"
-
-#define streq(a, b) (strcmp((a), (b)) == 0)
-#define strneq(a, b, n) (strncmp((a), (b), (n)) == 0)
-
-#define streqi(a, b) (strcasecmp(a, b) == 0)
-#define strneqi(a, b, n) (strncasecmp(a, b, n) == 0)
-
-static inline bool ascii_isdigit(char a) {
-        return a >= '0' && a <= '9';
-}
-
-static inline bool ascii_isalpha(char a) {
-        return (a >= 'a' && a <= 'z') || (a >= 'A' && a <= 'Z');
-}
+#include "string-util.h"
 
 #define UNUSED __attribute__((unused))
 
@@ -70,20 +57,6 @@ static inline void *malloc0_array(size_t base_size, size_t element_size, size_t 
         return malloc0(total_size);
 }
 
-static inline char *strcat_dup(const char *a, const char *b) {
-        size_t a_len = strlen(a);
-        size_t b_len = strlen(b);
-        size_t len = a_len + b_len + 1;
-        char *res = malloc(len);
-        if (res) {
-                memcpy(res, a, a_len);
-                memcpy(res + a_len, b, b_len);
-                res[a_len + b_len] = 0;
-        }
-
-        return res;
-}
-
 #define _cleanup_(x) __attribute__((__cleanup__(x)))
 #define _cleanup_free_ _cleanup_(freep)
 #define _cleanup_fd_ _cleanup_(closep)
@@ -104,5 +77,3 @@ static inline char *strcat_dup(const char *a, const char *b) {
 #define _cleanup_sd_bus_slot_ _cleanup_(sd_bus_slot_unrefp)
 #define _cleanup_sd_bus_message_ _cleanup_(sd_bus_message_unrefp)
 #define _cleanup_sd_bus_error_ _cleanup_(sd_bus_error_free)
-
-bool copy_str(char **p, const char *s);
