+++ src/libhirte/meson.build
@@ -7,7 +7,6 @@ libhirte_src = [
     'bus/utils.h',
     'common/cfg.c',
     'common/cfg.h',
-    'common/common.c',
     'common/common.h',
     'common/event-util.c',
     'common/event-util.h',
